Load unit shaders from xml, ally/enemy shaders for CloakType & send tangents to the shader for meshes with normal maps.
Move ShaderSet from model_renderer_gl.h/cpp to shader.h/cpp, rename it UnitShaderSet and made changes so that each 'set' is loaded from a single xml file.
Re-jig CloakType, can now have an 'ally-shader' (for the players own and allied cloaked units) and an 'enemy-shader' (for use on cloaked enemy units that have been detected).
Tangents are now sent to the shader (as a vertex attribute) for meshes with normal maps.