Re-add float distance to PosData to correct out-of-order results.
Converting PosData to fixed breaks qsort and causes
PosCircularIteratorFactory to generate out-of-order data, causing every
use of PosCircularIterator to be out of order.  Re-adding a floating
point distance value fixes this.

Also, I've replaced the call to qsort with std::sort, as I've recently
learned that it's actually faster because it can inline the compare
function instead of calling it by function pointer.